# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MentorAlm API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List published services
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a single service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /services/{id}/slots:
    get:
      operationId: getServiceSlots
      tags: [services]
      summary: Get available slots for a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
...
  /admin/services:
    post:
      operationId: createAdminService
      tags: [admin]
      summary: Create a service
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceInput"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /admin/services/{id}:
    patch:
      operationId: updateAdminService
      tags: [admin]
      summary: Update a service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceInput"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteAdminService
      tags: [admin]
      summary: Delete a service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Service:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        shortDesc:
          type: string
        fullDesc:
          type: string
        included:
          type: string
        category:
          type: string
        duration:
          type: string
        price:
          type: integer
        counsellorName:
          type: string
        counsellorBio:
          type: string
        thumbnailUrl:
          type: string
          nullable: true
        status:
          type: string
        slots:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - shortDesc
        - fullDesc
        - included
        - category
        - duration
        - price
        - counsellorName
        - counsellorBio
        - status
        - slots
        - createdAt
    CreateServiceInput:
      type: object
      properties:
        title:
          type: string
        shortDesc:
          type: string
        fullDesc:
          type: string
        included:
          type: string
        category:
          type: string
        duration:
          type: string
        price:
          type: integer
        counsellorName:
          type: string
        counsellorBio:
          type: string
        thumbnailUrl:
          type: string
        status:
          type: string
        slots:
          type: string
      required:
        - title
        - shortDesc
        - fullDesc
        - included
        - category
        - duration
        - price
        - counsellorName
        - counsellorBio
    UpdateServiceInput:
      type: object
      properties:
        title:
          type: string
        shortDesc:
          type: string
        fullDesc:
          type: string
        included:
          type: string
        category:
          type: string
        duration:
          type: string
        price:
          type: integer
        counsellorName:
          type: string
        counsellorBio:
          type: string
        thumbnailUrl:
          type: string
        status:
          type: string
        slots:
          type: string
...